Preserve network interfaces order of vnfd template
The Python dictionary behavior of vnfd body does not guarantee the
network interfaces to be fetched in same order as specified in template.
This fix uses the third party python package tosca-parser to preserve
the order. This will ensure the interfaces come up in the same sequence
as specified in template.

Closes-Bug: #1504687
Change-Id: Ieddd2ba16328869cef901d94bb53dc9566c250df