Fix mgmt_url 'NoneType' error can't be decoded
This fixes the "TypeError: <type 'NoneType'> can't be decoded"
stack trace that is dumped when VNF creation fails and
mgmt_url is None.

VNf should not be added to monitoring if it in ERROR state and
mgmt_url is None. The actual error is masked when 'None' mgmt_url
is decoded which can mislead the user.

Change-Id: I164bc03a78b9d0dd533506df09e5c267f6855ecb
Closes-Bug: #1524524