db schema improvements: blob to json text
Tacker does consider auth, placement_attr and similar
Objects as python dict but its stored as blob object
in db, which increases the storage space as well while
Debugging, these values can't be understandable as
its in binary format. So this patch stores all these
Objects in json string format, which enables readability
and reduces the size required to stored the object.

Closes-bug: #1587579

Change-Id: Idaee81028d40f60b70db12c5ae6ba78cb19f1992