Initial api-ref skeleton for Tacker
This patch is the first step to bring the NFV Orchestration
API Reference to the repository.
It is a part of the effort described here [1]:

[1] https://wiki.openstack.org/wiki/Documentation/Migrate#API_Reference_Plan

Once this skeleton patch is merged, an api-ref job can be added to
the gate. The gate job makes it much easier to review API Reference
contents, which will be brought by follow-on patches.

Change-Id: I7cc587f2c7b3c546be9cd11bad7e3b4eeacad171
Co-Authored-By: Naoya Harada <n-harada@vx.jp.nec.com>