Avoid Forcing the Translation of Translatable Variables
Translation can also be delayed for variables that potentially contain
translatable objects such as exceptions.
Whenever possible translation should not be forced by use of six.text_type()
on a message being used with a format string.
This patch allows the translation of the translatable replacement text to be
delayed until the message is translated.
http://docs.openstack.org/developer/oslo.i18n/guidelines.html#avoid-forcing-the-translation-of-translatable-variables

Change-Id: I5e259368cbf135116bb3b631835227caa1825894