Refactor the source code of infra_driver/openstack
In order to support Senlin driver for vnf auto-scaling management,
need to add auto-scaling backend switch function. But the openstack
infra_driver's source codes are using too many inner methods, that
is difficult to read, not feasible to change, and also difficult
to test. This patch refactors the source code for the first step.

* Move template translation to a separated module
* Move HeatClient class to a separated module

Change-Id: I944b3d530dcfc07d59a5efabe5cfbf788d40111a