Refactor: separating backend actions out from monitor
Currently, backend actions are tightly coupled with monitor module.
This is like a barrier when we have a various actions in the future.
Also, it will be good for separating monitoring-related features out
from tacker-server. This patch also proposes a policy action framework so
that VNF operators can easily add their own actions.

Change-Id: I6ad163f7435c3b778810f96c506c77298be0c53d
Closes-bug: #1582446