Disable debug log of ping command in vim ping action
The ping action is in a tight loop, if debug level is enabled on
tacker.agent.linux.utils, the mistral executor will take up too
much CPU resource and IO to do it. A debuglog flag argument is
introduced to utils.execute method to disable the debug log
even if the logging infrastructure is at debug level.

Change-Id: I0c84ac226edbd23afbfa258993eb3ef94d48d758
Implements: blueprint refactor-vim-monitor