[Bug-Fix] Update VNFFG leaves db in error state
Currently in tacker there is no support for updating a chain or
a classifier. So if anyone was trying to update a chain through
the vnffg-update command of the tackerclient a NotImplementedError
was raising and that was leaving the whole tacker ecosystem in an error
state.

That was happening because some objects in the database was in an status
ERROR and some other in a status PENDING_UPDATE.

With this patch we are setting all the objects in the database in a
ERROR status therefore the user can delete the failed VNFFG without
erasing the whole tacker database.

Closes-Bug: #1724809

Change-Id: I2c399a33055ccce41be941eca0bee52508399fd7
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>