Implement the vnf-mapping update functionality
Although tacker client side supports the update of the VNFs which
are currently used in an existing VNFFG this functionality is not
implemented at the Tacker server side.

Through this patch we are trying to address exactly that problem.
Also this patch is the first step for the implementation of this:

Implements: blueprint update-vnffg

Change-Id: I147e9f8a52f3e307767cd855492eb9bb5867cf29
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>