Remove setting of version/release from releasenotes
Release notes are version independent, so remove version/release
values. We've found that projects now require the service package
to be installed in order to build release notes, and this is
entirely due to the current convention of pulling in the version
information. Release notes should not need installation in order
to build, so this unnecessary version setting needs to be removed.
This is needed for new release notes publishing, see
I56909152975f731a9d2c21b2825b972195e48ee8 and the discussion starting at
http://lists.openstack.org/pipermail/openstack-dev/2017-November/124480.html .

Change-Id: I32dd5453b4d05e46cfd1cddfc5d2f885ce53da83