replace the class path uuid based stack names with vnf names
Tacker uses the class path+file name+uuid as stack names to
uniquely identify the stacks for vnf create and vnf respawn workflow.
Tacker now mandates VNF names to be provided in API layer and will be
made unique. It is useful to base stack names based on vnf names for
debugging purposes. This affects the naming of nova instances as well.

Change-Id: If043ad8c6c356f3fa1d2f5648d3d861ff5006602
Closes-Bug: #1590215