Add reusing feature when creating VNFFG.
Currently, when a VNFFG is created, Tacker also create port-pair,
port-pair-group, port-chain and flow-classifier. There are no
reusing ability in this approach.

This patch will check the port-id of a VNF is using in other VNFFG
or not. If it is used in a VNFFG, the existing port-pair-group will
be added to the new port-chain.

Therefore, we can create 2 VNFFGs that can go through one VNF.

This patch also clean up network resource such as port pair, port
pair group and flow classifier if there are failures.

Change-Id: I648693103dd7eab7a4b10cde89f11766a9232e8d
Closes-Bug: #1746721