Delete VNF should fail with 409 error
If user deletes VNF in "PENDING_CREATE" status, VNF is deleted from
tacker but the resources created by VIM driver are not cleaned up
properly. This patch doesn't allow VNF to be deleted if the status is
in "PENDING_CREATE" status. It will raise 409 error.

APIImpact
Return 409 error instead of 200 when VNF is deleted in
"PENDING_CREATE" status.

Change-Id: I2de9c3dce5a4b3795119e2550aa8acea68940c45
Closes-Bug: #1798726