Delete remaining VNFD in functinal test
This patch resolves an issue where if a functional test fails
in the middle of a run, VNFD remain and the next functional test fails.

This problem is that if the function test fails in the middle,
the VNFD created in the function test remains without being deleted,
so if the same VNFD name is used in the next function test executed,
the VNFD creation will fail.

Therefore, the following changes have been made.

* Changed the timing of calling addCleanup()
* Made VNFD creation a common function

This patch fixes the following bugs, but VNF does not.

Change-Id: I8a932fbe3379604b0c3173d2c06e6ed7abb8beaf
Partial-Bug: #1771494