Add unique constraint on vnfd_id column
In Vnf LCM, when user will call create vnf instance API, vnfdId
will be passed in the CreateVnfRequest as per SOL002,
section 5.5.2.3 [1]. Based on this vnfdId, it will look for the
vnf package to load the csar package at the time of instantiating vnf
instance. So it's important that vnfdId is unique in the
``vnf_package_vnfd`` dbb table.

Added ``vnfd_id`` and ``deleted`` columns as unique key in
``vnf_package_vnfd`` db table.

[1] : https://www.etsi.org/deliver/etsi_gs/NFV-SOL/001_099/002/02.06.01_60/gs_nfv-sol002v020601p.pdf

Change-Id: I2da2acf256bec307b606c5a152c0d47af24ad525
Blueprint: support-etsi-nfv-specs