Revert vim shell script file mode from 644 to 755
In patch [1], the mode of tools/test-setup-default-vim.sh shell
script file was changed from 755 to 644 because of which it's fails to
run the script which registers a default vim 'VIM0' needed to run
for some of the functional tests.

This patch reverts back the mode of test-setup-default-vim.sh
from 644 to 755 and logs meaningful error message in case
the script file doesn't exists or it doesn't have execute permission.

[1] : https://review.opendev.org/#/c/674761

Change-Id: I4651bdd89da12720416e630173f80cc788911307
Closes-Bug: #1874007