Add UT instead of deleting FT related with Mistral
This patch deletes the FT related with Mistral for the following reasons.
- Other OpenStack-PJs (e.g. nova and heat) that have Action definition in
mistral([1]) do not implement FT in each PJ.
- These functional tests are carried out in the mistral-tempest-plugin.
However, at the moment, only core PJs (e.g. nova and heat) are targeted in [2],
and it's necessary to aim to add tacker here in the future.

In this patch, instead of deleting FT, the viewpoint of unittest is added to
confirm that mistral-client is called as expected.
The code is modified so that the anchor notation is not used
because not supported by mistral in [3].

[1]: https://opendev.org/openstack/mistral-extra/src/branch/master/mistral_extra/actions/openstack/mapping.json
[2]: https://opendev.org/openstack/mistral-tempest-plugin/src/branch/master/mistral_tempest_tests/tests/scenario/engine/actions/v2/test_openstack_actions.py
[3]: https://opendev.org/openstack/mistral/commit/eac23d9e774f658f9d4743c99aa2743eb104c3f9

Closes-Bug: #1878309
Change-Id: Ie8eab1b262c15f7b53796a287b333e60ffee1c31