Unexpected update in setup script for functional test
There are some placeholder files for running functional test in zuul.
The contents of the files, such as `auth_url` of the following files,
are updated while running devstack `stack.sh`.

* tacker/tests/etc/samples/local-vim.yaml
* tools/test-setup-default-vim.sh

These Git management files including the above, should not be
modified. Therefore, I will be fix as follow.

The need files to be change for Zuul are deploy by Ansible,
so the deploy process is run with Ansible instead of the
`devstack/lib/tacker` script.

Allow `tools/test-setup-default-vim.sh` script to be executed
by itself, so the git management files aren't updated by running
devstack `stack.sh`.

Also, as other improvements, replace the file path included in
Ansible role to definition value.

Change-Id: Iad88c7adfe56e926ee0324d94787577ba066989b
Closes-Bug: #1879303