Fix failed test for creating vnf image
The number of getting image URL is expected as `2`, but notactually
in test_pre_instantiation_vnf_image_with_file(). The value can be `1`
or `2` for `tox -e py36` or `tox -e lower-constraints`. It might be
because of a difference of versions of modules imported in
the test cases.

This update is just a quick fixing to avoid the issue. So, it could be
happened again possibly if the versions of modules are changed.

Closes-Bug: #1885523

Change-Id: I302607f636d6d96baa544ccad600bb21fdf18da2
Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>