Fix: Instantiation fails with single VNFD
When executing "Instantiate a VNF instance" with single VNFD,
it fails with an error.
This is because in instantiation, it's not possible to consider
that VNFD is single, and an error will occur when trying to get
the information of VNFD imported on the multiple premise.

In this patch, if the VNFD is single, that is,
if the imported VNFD doesn't exist, fix to skipping
the subsequent VNFD information getting process.
In unittest, I found an error in the existing single VNFD,
so correct it as well.

Change-Id: I9e664ec156bdcf3a95ecc1c6a0b582604f7f0335
Closes-Bug: #1886634