Running the VNF LCM instantiate using the sample vnf package fails
The CSAR file for VNF package registered as a sample fails to run
the VNF LCM instantiate. The cause of this issue is typo and invalid
definition for the VNFD. I think that the csar zip file for uploaded VNF
package should be able to run the VNF LCM instantiate normally.
Therefore, fix typo and invalid definition for the VNFD in the CSAR file.

Also, fix to the following settings that has a dependent on environment
because the current VNFD can only be run in the limited environment.

* Change hash algorithm from sha-256 to sha-512 according
  the Glance default hash algorithm
* Change from 8192 MB to 256 MB in the `min_ram` size of
  the VirtualStorage to work even in a low specification environment
* Delete `CP1.properties.vnic_type` to work even under a virtual
  environment

Change-Id: Icccaa523d33115e8b7ec0a6fd7e9c5f177d0a363
Closes-Bug: #1886124
(cherry picked from commit b397d01a3e4fb77809c8c48e78eebe8cf2a3fcf4)