Fix for package error during vnflcm creation
Execution "openstack vnflcm create" raises some package errors.
In order to avoid them, the exception process in vnflcm creation
is fixed.

Closes-bug: #1898963
Closes-bug: #1899112
Closes-bug: #1899134

Additional Information:
---------------------------------------
- vnf_lcm_subscription.id was decoded. The code we developed in
advance was created on the premise of varbinary and there was an
unmatch, therefore, at first, the DB type was matched. To modify
the code correctly, id decode was removed.
- Spec (Support LCM notifications for VNF based on ETSI NFV-SOL
specification)
There was a discrepancy with the content and I fixed the registered
migration, but the procedure was incorrect, so it was restored.
- Spec (Support LCM notifications for VNF based on ETSI NFV-SOL
specification)
I found that some columns were not added to the DB due to a mismatch
with the contents, so I am adding a new migration file. Also, since
the type definition was different for the already registered columns,
it was also corrected.

Change-Id: I0ee9d7e1e0c4ea26dd5894ad4fca302fe9bfd4d1