Temporarily remove Ceilometer installation
Ceilometer installation seems to be the cause of POST_FAILURE,
so this patch remove it temporarily.

Related error
https://zuul.opendev.org/t/openstack/build/5abe3c88e63249d0a6ad5bd764a9d604

Ceilometer installation was removed before, because Gnocchi
installation was failed. [1]
After above error was resolved, we reverted Ceilometer
installation. [2]
However, the tests using Ceilometer do not succeed as mentioned
at the beginning of this commit message.

[1] https://review.opendev.org/#/c/751965/
[2] https://review.opendev.org/#/c/754882/

Change-Id: I5c6ddd31abd58fa2941a3a1919f09f473dfd836e