Change tosca-parser and heat-translator install
With devstack, tosca-parser and heat-translator of stable version are
installed via pip as listed in `requirements.txt`. However, we usually
use devsatck for deploying development env, and it's recommended to get
the latest division sometimes. On the other hand, there are not so many
cases using stable ones actually.

This update includes several tricks as following, but most of them are
no need if the two packages have its own devstack script. So, we should
remove the tricks after that as described in TODOs added in the update.

NOTE: As described below, we should update upper-constraints.txt as
following example and upload it to tacker's repo everytime [1] is
updated for the change.

  $ bash devstack/lib/download_upper_consts

* Install packages from cloned git repos in `/opt/stack/` as similar
  to other packages.

* Comment out entries of the two packages in `requirements.txt` and
  `lower-constraints.txt` because no need to install them from here,
  but activate it if you install stable packages.

* To pass tox jobs, add files below.

  * upper-constraints.txt: The two packages are removed from official
    constraints file [1] to skip checking version of stable. This file
    is refreshed everytime running `stack.sh` to catch up the official.

  * requirements-extra.txt: Install the packages from git repo. Without
    separating it from `requirements.txt`, `tox -e lower-constraints`
    is failed.

* Update params in `tox.ini` to activate the modifications above.

[1] https://releases.openstack.org/constraints/upper/master

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: I776137215ea06c8ac54612d13a663349c2e7db7d