Change data size of instance_id in VNF table
Need to change data size of `instance_id` in the VNF table to
String(12800) from String(64).
We don't have an appropriate ID for instance_id for CNF case, so we use
the object information instead as a workaround.
If it leave the definition of String(64) as it is, a "Data too long for
column" error will occur during CNF operation, because instance_id
store kubernetes object information such as following:
----
'instance_id': "
  {'namespace': 'default', 'name': 'vdu1',
   'apiVersion': 'apps/v1', 'kind': 'Deployment'},
  {'namespace': 'default', 'name': 'pv1',
   'apiVersion': 'core/v1', 'kind': 'PersistentVolume'},
  ... (**Contains the deployed kubernetes resources)
  {...}"
----

Reason of changing string size to 12800:
Resize to 12800 so that it can store approximately 100 resources when
it requires 128 characters per resource.
This 12800 characters is under the limit of 65535 bytes per entry,
even for the maximum length character set (4 bytes) [1].
(instance_id's size (12800 * 4) + other total size(3044) < 65535)
`other total size` means the capacity used other than instance_id in
the VNF table.
[1] https://dev.mysql.com/doc/refman/8.0/en/charset-charsets.html

Closes-Bug: 1909011
Change-Id: I50db6f86cd8c164efeab2f0ebbe8ec0258a5ef1d