Fix some content of VNFD in UT and FT
Recently tosca-parser has started to validate the required properties
of policies [1]. In current implementation, this caused errors in
testing with VNFD where the policy definition was incorrect.

[1] https://review.opendev.org/c/openstack/tosca-parser/+/763144

This patch fixed the policy type in the VNFD which errored with UT and
FT, so that the VNFD can pass the enhanced policy type validation
processing in tosca-parser.

Closes-Bug: #1909312
Change-Id: I0b655d6ebce7c4dacb9cc9c0526aa702d1ac37db