Workaround for the ControllerRevision FT
This patch is provisional workaround for bug [1].

I think possible ways to fix the bug are using the kubernetes
python client v11.0.0. However, as a other issue, the
ControllerRevision resources is not delete and remain. The
problem of resources remaining cannot be solved immediately.
Therefore, add a skip decorator to skip the Functional Test of
ControllerRevision as a workaround.

The added skip decorator will be removed after this issue
completely resolved.

[1] https://bugs.launchpad.net/tacker/+bug/1910327

Partial-Bug: #1910327
Change-Id: I6bedc5c8c712c9cf7d95ae5cb7384c4dea187108