Fix bug only last vdu will be saved to nested_tpl
When instantiating multiple VDU using `OS::Heat::AutoScalingGroup`
type in vnfd with SOL format, an error about not being able to
find the specified type occurred.

In current `for loop` of update_nested_scaling_resources(),
only the last data will be stored into the nested_tpl[],
the other data will be lost.

This modification changed the logic of the `for loop`,
it will store all types of data configuration into nested_tpl[]
according to the `aspect id`.

Closes-Bug: 1917314
Change-Id: I40eb711b7d0e8e2cf61f0210151be60febd08407