Temporary revert to install ht and tp with pip
This update is to revert a change for tosca-parser and heat-translator
installation [1]. It works fine for devstack for Tacker, but doesn't
for some CI tests in kolla. It fails for a "package cannot be found"
error.

The reason of [1] is introduced is because we need to make a change and
it can be used from Tacker quickly. We cannot wait pip packages updated.
However, the latest versions are released recently in which all updates
required from Tacker are included. So, it's OK to revert [1] for a
while.

If this reverting does not affect to kolla CI after merged, we should
cancel the change and make install from git repo again.

[1] commit 500d2eb7 "Change tosca-parser and heat-translator install"

Signed-off-by: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Ie2d994b053e909ab10a99307793c795628482e2b