Support Retry of Change External Connectivity
This feature will enable the client to use this API to initiate
retrying Change VNF External VNF Connectivity operation when it
fails.

It is based on the ETSI NFV specification that retry can be used
when a VNF lifecycle operation is in a FAILED_TEMP state.

Implements: blueprint support-change-external-connectivity
Spec: https://specs.openstack.org/openstack/tacker-specs/specs/wallaby/support-change-external-VNF-connectivity-operation.html
Change-Id: Ia1d907922cdb701586f15c4840848fa8402ded3b