Enable neutron plugin for Octavia
This patch cleary enables a `neutron plugin` for the functional test
of `tacker-functional-devstack-multinode-sol-kubernetes` in zuul.

[1] is fixed in the Octavia on the assumption that using a `octavia
create_network_interface_device` function[2] of the Neutron.
According to [1], if using the Octavia, it seems that the
`neutron plugin` needs to be clearly define.

Also currentry the functional test of
`tacker-functional-devstack-multinode-sol-kubernetes` is not became
`-1` status on zuul when zuul's result even is `RETRY_LIMIT` because
 job's voting is `non-vorting`. So this patch changes from
`non-voting` to `voting` in zuul's job results.

[1] https://github.com/openstack/octavia/commit/651f095a818c9a30441ed8df9a39cf726b6e3f5d
[2] https://github.com/openstack/neutron/blob/master/devstack/lib/octavia#L7-L28

Closes-Bug: 1919396
Change-Id: I999a2c7c06528c7470cec2161bdd361a67a8da6f