Fix a regex of naming rule for a Pod of a Deployment
This patch is fix a regex of naming rule for a Pod of a Deployment.

The naming rule of Pod in deployment is
"(deployment name)-(pod template hash)-(5 charactors)".

Currently, "pod template hash" is a regex with a fixed 10 characters.
However, it may be less than 10 characters, this case is fail in
"scale_wait" of CNF scale.

The "pod template hash" string is generated from 32 bit hash. This
may be from 1 to 10 caracters but not sure the lower limit from the
source code of Kubernetes. Temporarily, this is set to be "{1,10}"
but may need to be fixed.

Closes-Bug: #1921059
Change-Id: I49f6cbe032f35f435f6c9ee69ccb8596cccdb326