Addition of location header field in response
For below VNFInstance's operations "Location header" will be present
1) heal
2) instantiate
3) terminate

Response header will contain the "Location" HTTP header field.
And this "Location" HTTP header will contain "vnfLcmOpOccId".

This patch changes notification parameter based on
SOL013  7.1.4 Type: NotificationLink.

As per SOL003, 5.5.2.17, VnfLcmOperationOccurrenceNotification
represents a VNF lifecycle management operation occurrence
notification, which informs the receiver of changes in the
VNF lifecycle caused by a VNF LCM operation occurrence.

This notification shall be triggered by the VNFM,
when there is a change in the state of a VNF LCM operation
occurrence that changes the VNF lifecycle, including:
• Instantiation of the VNF
• Scaling of the VNF instance (including auto-scaling)
• Healing of the VNF
and others.

URI of a resource referenced from a notification
should be an absolute URI if APIRoot is known.

Co-Authored-By: Sheel Rana <ranasheel2000@gmail.com>

Closes-Bug: #1915107
Closes-Bug: #1911625
Change-Id: I3a325634f88b9521d7fa2f524076934df315742d