Fix UT and FT VNFD yaml because lack of properties
Since validation of VNFD is more strict, `target` attribute should be
under `properties` for `tosca.policies.tacker.Scaling`, or failed to run
unit and functional tests. This update is to correct VNFDs.

This patch also updates version of heat-translator to the latest
`2.3.0` in `requirements.txt` and `lower-constraints.txt` to ensure it
uses the version for the change.

Co-Authored-By: Yasufumi Ogawa <yasufum.o@gmail.com>
Change-Id: Iae2b25abe53eda483e811171306d3d375a7e1687
Closes-Bug: #1921996